pyproject.toml
[project]
name = "my-server-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "atlassian-python-api>=4.0.7",
    "fastmcp>=2.13.0.2",
    "mcp[cli]>=1.21.0",
]
